Triangle Analyzer Console Application (C#)
A C# .NET Framework Console Application that takes three user inputs and determines if the inputs create a triangle. If they do, then it will analyze the inputs and return either as an equaliteral, isosceles, or scalene triangle. There are also Unit tests that run with NUnit within the .tests folder.